requirements.txt
Flask
joblib
scikit-learn
newspaper3k
wheel==0.45.1
gunicorn
PyMuPDF
